backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
langchain
langchain-nvidia-ai-endpoints
langchain-core
